+++ iMe.Api/Factory/ClientFactory.cs
@@ -4,19 +4,19 @@ using Microsoft.Practices.Unity;
 
 namespace iMe.Factory
 {
-    public class ClientFactory : IClientFactory
+    public class ClientFactory 
     {
-        public ISocialNetworkClient GetClient(string clientType)
-        {
-            var container = UnityConfig.GetConfiguredContainer();
+        //public ISocialNetworkClient GetClient(string clientType)
+        //{
+        //    var container = UnityConfig.GetConfiguredContainer();
 
-            switch (clientType.ToLowerInvariant())
-            {
-                case "twitter":
-                    return container.Resolve<ISocialNetworkClient>("twitter");
-                default:
-                    return container.Resolve<ISocialNetworkClient>();
-            }
-        }
+        //    switch (clientType.ToLowerInvariant())
+        //    {
+        //        case "twitter":
+        //            return container.Resolve<ISocialNetworkClient>("twitter");
+        //        default:
+        //            return container.Resolve<ISocialNetworkClient>();
+        //    }
+        //}
     }
 }
