+++ iMe.Bootstrapper/App_Start/UnityMvcActivator.cs
@@ -1,0 +1,34 @@
+using System.Linq;
+using System.Web.Mvc;
+using Microsoft.Practices.Unity.Mvc;
+
+[assembly: WebActivatorEx.PreApplicationStartMethod(typeof(iMe.Bootstrapper.App_Start.UnityWebActivator), "Start")]
+[assembly: WebActivatorEx.ApplicationShutdownMethod(typeof(iMe.Bootstrapper.App_Start.UnityWebActivator), "Shutdown")]
+
+namespace iMe.Bootstrapper.App_Start
+{
+    /// <summary>Provides the bootstrapping for integrating Unity with ASP.NET MVC.</summary>
+    public static class UnityWebActivator
+    {
+        /// <summary>Integrates Unity when the application starts.</summary>
+        public static void Start() 
+        {
+            var container = UnityConfig.GetConfiguredContainer();
+
+            FilterProviders.Providers.Remove(FilterProviders.Providers.OfType<FilterAttributeFilterProvider>().First());
+            FilterProviders.Providers.Add(new UnityFilterAttributeFilterProvider(container));
+
+            DependencyResolver.SetResolver(new UnityDependencyResolver(container));
+
+            // TODO: Uncomment if you want to use PerRequestLifetimeManager
+            // Microsoft.Web.Infrastructure.DynamicModuleHelper.DynamicModuleUtility.RegisterModule(typeof(UnityPerRequestHttpModule));
+        }
+
+        /// <summary>Disposes the Unity container when the application is shut down.</summary>
+        public static void Shutdown()
+        {
+            var container = UnityConfig.GetConfiguredContainer();
+            container.Dispose();
+        }
+    }
+}
