+++ iMe.Bootstrapper/packages.config
@@ -1,10 +1,16 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="CommonServiceLocator" version="1.3" targetFramework="net452" />
+  <package id="Microsoft.AspNet.Mvc" version="4.0.20710.0" targetFramework="net452" />
+  <package id="Microsoft.AspNet.Razor" version="2.0.20710.0" targetFramework="net452" />
   <package id="Microsoft.AspNet.WebApi.Client" version="5.2.3" targetFramework="net452" />
   <package id="Microsoft.AspNet.WebApi.Core" version="5.2.3" targetFramework="net452" />
   <package id="Microsoft.AspNet.WebApi.WebHost" version="5.2.3" targetFramework="net452" />
+  <package id="Microsoft.AspNet.WebPages" version="2.0.20710.0" targetFramework="net452" />
+  <package id="Microsoft.Web.Infrastructure" version="1.0.0.0" targetFramework="net452" />
   <package id="Newtonsoft.Json" version="6.0.4" targetFramework="net452" />
   <package id="Unity" version="4.0.1" targetFramework="net452" />
+  <package id="Unity.Mvc" version="4.0.1" targetFramework="net452" />
   <package id="Unity.WebAPI" version="5.2.3" targetFramework="net452" />
+  <package id="WebActivatorEx" version="2.0" targetFramework="net452" />
 </packages>
