+++ iMe.sln
@@ -23,6 +23,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Interfaces", "Interfaces", 
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Common", "Common", "{6927A778-5B06-448A-A705-C6BB0F7B26D6}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "iMe.Bootstrapper", "iMe.Bootstrapper\iMe.Bootstrapper.csproj", "{10B31D72-2640-44ED-8835-3B31D5ACBD32}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -49,6 +51,10 @@ Global
 		{4CA69625-14F1-4433-9BE1-0A3E7093E2B3}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4CA69625-14F1-4433-9BE1-0A3E7093E2B3}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{4CA69625-14F1-4433-9BE1-0A3E7093E2B3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{10B31D72-2640-44ED-8835-3B31D5ACBD32}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{10B31D72-2640-44ED-8835-3B31D5ACBD32}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{10B31D72-2640-44ED-8835-3B31D5ACBD32}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{10B31D72-2640-44ED-8835-3B31D5ACBD32}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -61,5 +67,6 @@ Global
 		{823D7955-9F8E-4601-9AE2-946DE171A06A} = {4F127CC6-0ACE-4103-981D-B07A21A5C3C5}
 		{4CA69625-14F1-4433-9BE1-0A3E7093E2B3} = {A8AEB428-F551-449C-BD04-C83B9D80FE51}
 		{A8AEB428-F551-449C-BD04-C83B9D80FE51} = {874738DB-AFAA-497E-B2DF-6AC99596CBE1}
+		{10B31D72-2640-44ED-8835-3B31D5ACBD32} = {6927A778-5B06-448A-A705-C6BB0F7B26D6}
 	EndGlobalSection
 EndGlobal
